Sort through the data and add it to each country
Add a variable called `happyData` that gets filled with the data for each country, using the country code as the key and having the value be an array where each index is a years worth of data from the combined CSV.

Not every country has an entry or an entry for every year so some arrays will be shorter then the maximum of 5 and countries with no information will have a value of `null`.

The information for each country is stored in the key `info` in each county's object.